+++ src/JirungEngine/MathModule/Point/Point.hpp
@@ -3,20 +3,20 @@
 
 namespace JirungEngine {
     class Point {
-        public: int x;
-        public: int y;
-        public: int z;
+        public: float x;
+        public: float y;
+        public: float z;
 
     public:
         Point();
-        Point(int x, int y, int z = 0);
+        Point(float x, float y, float z = 0);
         Point(const Point& point);
 
         Point operator=(const Point& point);
         Point operator+(const Point& point);
         Point operator-(const Point& point);
-        Point operator*(int n);
-        Point operator/(int n);
+        Point operator*(float n);
+        Point operator/(float n);
     };
 }
